refactor: crdt lib를 common workspace로 이동
