--- packages/server/src/lib/crdt/identifier.ts
@@ -1,10 +1,0 @@
-class Identifier {
-  clock: number;
-  client: number;
-  constructor(clock: number, client: number) {
-    this.clock = clock;
-    this.client = client;
-  }
-}
-
-export default Identifier;
