+++ packages/server/src/modules/events/crdt/crdt.gateway.service.ts
@@ -2,8 +2,8 @@ import { Injectable, Logger } from '@nestjs/common';
 import { Server, Socket } from 'socket.io';
 import { SOCKET_EVENT } from '@collab-note/common';
 import { RemoteDeleteDto, RemoteInsertDto, RemoteUpdateDto } from './dto';
-import CRDT from '~/lib/crdt/crdt';
-import LinkedList from '~/lib/crdt/linkedList';
+import { CRDT, LinkedList } from '@collab-note/common';
+
 @Injectable()
 export class CrdtGatewayService {
   private server: Server;
