+++ packages/server/src/modules/events/crdt/dto/remote-insert.dto.ts
@@ -1,4 +1,4 @@
-import Node from '~/lib/crdt/node';
+import { Node } from '@collab-note/common';
 
 export class RemoteInsertDto {
   id: string;
